requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
click==8.1.7
Flask==3.1.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
pillow==11.0.0
python-docx==1.1.2
python-pptx==1.0.2
soupsieve==2.6
typing_extensions==4.12.2
Werkzeug==3.1.3
Whoosh==2.7.4
XlsxWriter==3.2.0
zipp==3.21.0
